re-apply dsp-related changes from r4533
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4569 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DSPCore/Src/DspIntArithmetic.cpp
@@ -33,7 +33,7 @@ void clr(const UDSPInstruction& opc)
 	u8 reg = (opc.hex >> 11) & 0x1;
 
 	dsp_set_long_acc(reg, 0);
-	Update_SR_Register64((s64)0);   // really?
+	Update_SR_Register64(0);
 	zeroWriteBackLog();
 }
